Removes qtst dependence on signed type
For CES type A-2 controllers, the 16th bit of the control status
register contains the Q status bit. Tests of this bit used to assume
that this bit was a sign bit in a signed integer by comparing the
total 16-bit word against zero. However, the camread16 macro returns
an unsigned number and this qtst always returned a value greater than
or equal to zero (i.e. qtst is always returning true). This is fixed.

Should verify that the BiRa controller data layout returned is identical to the
CES controller.

This fixes a bug introduced in rev. 2832.