Issue #1920 Adds control to select global mode register bits in VMUSB
MockVMUSB is added to test that the driver works at the 
software level. All tests pass for the new functionality of the vmusb.
New defaults for the global mode register are provided that seek to 
avoid running the VMUSB in modes that are less tested and have found
to be buggy (i.e. span buffer mode). 
The default global mode setting is now: 
- no spanning buffers, 
- no mixed buffers,
- no force scaler dump, 
- bus request level = 4, 
- no optional header, 
- align 16.