Refactors the CXLM base class
- To support code sharing for firmware loads in the
slow controls server and in the readout controls,
CXLM is put into the XLM namespace and all generic
functionality is factored out of the class scope
into the namespace scope. The various supported 
derived classes of CXLM are updated to reflect the
now qualified CXLM name. This is accomplished with
a "using namespace XLM".

- Slow controls are provided to the XLM to support
  loading firmware when an init is called.