Major refactoring of VMUSB classes
The following methods of the CVMUSB class have been made purely
virtual:
  - readRegister
  - writeRegister
  - executeList
  - loadList
  - usbRead

The support for interactive variable block reads has been removed in all of 
the classes because it did not have a valid use case and is accomplished in
a more straightforward manner through the CVMUSBReadoutList class.

The writeActionRegister no longer has special treatment and is treated in the same
way any other register is treated.

Duplicate code for bit encoding in stacks has been placed into the base class.

CMockVMUSB no longer implements its own vme commands and register operations, but
instead implements the bare minimum interface. Doing so actually makes is a 100%
drop-in replacement for a CVMUSB-derived object whereas in the past it only supportd
a subset of functionality. Note that the changes are backwards compatible and did not
break any unit tests that utilized it.

CVMUSBusb and CVMUSBEthernet now have access to the shadow register maintained
by the CVMUSB base class.

CVMUSBEthernet refactoring has not yet been completed because testing of its
original functionality needs to be accomplished first.