Issue #1921 support for optional header and fix for buffer allocation in COutputThread
The output thread now checks the shadow register of the vmusb and skips the optional
header if it is being used. A check is also performed to make sure that it makes sense.
If it disagrees, then it will print error messages. 

Also, for very large buffer sizes, the COutputThread::event() was reallocating buffers
incorrectly and causing seg faults. This fixes the issue.