Fixes end of run scaler dump in ScalerDisplay
The total scaler is actually a floating point number but the 
format command was looking for an integer. This caused an
exceptional return from the format command which broke the
rest of the dump. I updated the field of the formatting string
to expect a very large floating point number (14 digits) and 
zero digits right of the decimal. This is in line with the 
used to display the total scaler value, though a few digits
wider.