Adds Makefile generation for usb/common and also improves tcllib dependency resolution
Previously, if a user needed to build a local copy of tclplus in libtcl, they would
have to run ./configure with root permissions because of the make install of libtclplus
would need to write to directories owned by root (/usr/opt/nscldaq/...). This caused 
wasted time while building and installing nscldaq because all of the build code
created by ./configure would be owned by root, which seemed to influence dependency checking.
I don't know how. But this improves the situation by allowing configure to checks if the user
has already installed libtclplus in the prefix path before. If no local libtcl has been 
installed, it behaves as normal and does a complete local build and installation of 
whatever lives in libtcl.