Sprint impediments Task #5321 - TclBindings to CStateProgrsm
Allow for instantiation of the object with only 1 URI in which case a
CStateProgram is used as the API only.  If two URI's are used, an
CStateManager is instantiated and _its_ program api is used...except
for things which are cached and affect the state monitor
(the program parent directory).