Bug #6136 - Coincidence interval wiget of evb does not get used.
As indicatd in the bug text the prolem actually is that if the
event builder is set to be persistent, its control panel is useless
since the parameters it contains are only applied when the
event builder starts.

This commit provides simple code to know when the control panel
contents have changed.   If they have changed, the event builder is
then restarted at the next begin run.

Still needed:

*   Incorporate the new megawidget based control panel.
*   If the control panel contents are changed, and the event builder
    is persistent, warn the user that the event builder would need
    to be restarted at the next begin (giving them an opportunity
    to cancel the change.