Fixes BLT limitations caused by improper read back
To support reading back data sizes greater than 13 kilowords, it is
necessary to iteratively read out the VM-USB over and over again with
smaller buffer sizes until the full data payload is received. Doing so
allows for BLTs in interactive mode that can transfer up to 4 MB as
the VM-USB manual claims is possible. Also, the amount of data requested
to be read back during a writeRegister operation is changed from 32-bits
to 16-bits. This was essential to adhere to the proper functionality of
the VM-USB and also to satisfy the new strictures on the read back
portion of a transaction.